+++ CHANGELOG.md
@@ -31,6 +31,9 @@ Due to the removal of legacy code, there are a few breaking changes in this new 
 * Added support for multiple string tables.   
   [David Jennes](https://github.com/djbe)
   [#41](https://github.com/SwiftGen/templates/issues/41)
+* Storyboards now provide a `platform` identifier (iOS, macOS, tvOS, watchOS).  
+  [David Jennes](https://github.com/djbe)
+  [#45](https://github.com/SwiftGen/templates/issues/45)
 
 ### Internal Changes